Move directory and split file
